Move the Discord adapter into Services
