+++ Modix.sln
@@ -24,8 +24,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Modix", "Modix\Modix.csproj", "{C4BE4089-0983-464F-9D76-EE9BD2B708DD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Modix.Adapters.Discord", "Modix.Adapters.Discord\Modix.Adapters.Discord.csproj", "{3D15CCB1-7012-4357-A7F1-050084A1611F}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Modix.Services.Test", "Modix.Services.Test\Modix.Services.Test.csproj", "{D9BFC7C7-10B2-4D61-99EA-79805074CCD7}"
 EndProject
 Global
@@ -97,18 +95,6 @@ Global
 		{C4BE4089-0983-464F-9D76-EE9BD2B708DD}.Release|x64.Build.0 = Release|Any CPU
 		{C4BE4089-0983-464F-9D76-EE9BD2B708DD}.Release|x86.ActiveCfg = Release|Any CPU
 		{C4BE4089-0983-464F-9D76-EE9BD2B708DD}.Release|x86.Build.0 = Release|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Debug|x64.Build.0 = Debug|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Debug|x86.Build.0 = Debug|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Release|x64.ActiveCfg = Release|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Release|x64.Build.0 = Release|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Release|x86.ActiveCfg = Release|Any CPU
-		{3D15CCB1-7012-4357-A7F1-050084A1611F}.Release|x86.Build.0 = Release|Any CPU
 		{D9BFC7C7-10B2-4D61-99EA-79805074CCD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D9BFC7C7-10B2-4D61-99EA-79805074CCD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D9BFC7C7-10B2-4D61-99EA-79805074CCD7}.Debug|x64.ActiveCfg = Debug|Any CPU
