+++ Modix/Extensions/ServiceCollectionExtensions.cs
@@ -3,11 +3,11 @@ using Discord;
 using Discord.Commands;
 using Discord.WebSocket;
 using Modix;
-using Modix.Adapters.Discord;
 using Modix.Behaviors;
 using Modix.Data.Models.Core;
 using Modix.Data.Repositories;
 using Modix.Services;
+using Modix.Services.Adapters;
 using Modix.Services.AutoCodePaste;
 using Modix.Services.BehaviourConfiguration;
 using Modix.Services.CodePaste;
